chore: reorganize packages
I am moving the provisional queue to be explicitly under the queue strategies,
because it is always interacted with from there, and the provisional queue
is a strategy in and of itself that might need to be tweaked depending on
which queue strategy is in use. I am also adding the (missing) sort so
that earlier pod groups are submit first, and tweaking the docs slightly.
After this I will test running reservations.
