+++ kubernetes/pkg/fluxnetes/resources/resources.go
@@ -1,4 +1,4 @@
-package podspec
+package resources
 
 import (
 	"strings"
